Fixed code: -Renamed beans. -Moved config Jwt class to config package
